modified get territories

+++ src/pieces/territory.rs
@@ -10,7 +10,7 @@ impl Goban {
     pub fn get_strings_of_stones_without_liberties_wth_color(
         &self,
         color: Color,
-    ) -> impl Iterator<Item = GoStringPtr> + '_ {
+    ) -> impl Iterator<Item=GoStringPtr> + '_ {
         self.go_strings()
             .iter()
             .filter_map(|x| x.as_ref())
@@ -55,7 +55,7 @@ impl Goban {
     ///
     pub fn get_strings_from_stones(
         &self,
-        stones: impl Iterator<Item = Stone>,
+        stones: impl Iterator<Item=Stone>,
     ) -> Vec<HashSet<Stone>> {
         let mut groups_of_stones: Vec<HashSet<Stone>> = Default::default();
         for s in stones {
@@ -75,18 +75,18 @@ impl Goban {
     /// Returns (black territory,  white territory)
     ///
     #[inline]
-    pub fn calculate_territories(&self) -> (f32, f32) {
+    pub fn calculate_territories(&self) -> (u32, u32) {
         let (black_territory, white_territoty) = self.get_territories();
         (
-            black_territory.count() as f32,
-            white_territoty.count() as f32,
+            black_territory.count() as u32,
+            white_territoty.count() as u32,
         )
     }
 
     ///
     /// Get two iterators of empty stones.
     ///
-    pub fn get_territories(&self) -> (impl Iterator<Item = Stone>, impl Iterator<Item = Stone>) {
+    pub fn get_territories(&self) -> (impl Iterator<Item=Stone>, impl Iterator<Item=Stone>) {
         let empty_strings = self.get_strings_from_stones(self.get_stones_by_color(Color::None));
         let mut white_territory = Vec::new();
         let mut black_territory = Vec::new();
